Added javadoc to progress classes
Added javadoc comments to all function for progress and graph display, removed DaysUtil class
